customer_service_assistant_3agents/requirements.txt
sentence-transformers>=2.2.0
python-docx>=0.8.11
chromadb>=0.4.0
python-dotenv>=1.0.0
numpy>=1.21.0
torch>=1.9.0
transformers>=4.20.0
jupyter>=1.0.0
ipykernel>=6.0.0 